+++ src/utils/logs.ts
@@ -74,6 +74,7 @@ const logs = {
     SETUP_IN_START: '%s1 - Inside FrameElements start(), calling setup',
     SETUP_IN_CONSTRUCTOR: '%s1 - Inside FrameElements constructor, calling setup',
     COLLECT_ELEMET_START: '%s1 - %s2, Collect Element start',
+    COLLECT_CONTROLLER_START: '%s1 - %s2, Collect Controller start',
     REVEAL_ELEMENT_START: '%s1 - %s2, Reveal Element start',
     EMIT_COLLECT_FRAME_CONTROLLER_EVENT: '%s1 - Emit collect FrameController FRAME_READY event',
     EXECUTE_COLLECT_CONTROLLER_READY_CB: '%s1 - Executing collect FrameController FRAME_READY callback, set client metadata',
@@ -273,7 +274,9 @@ const logs = {
     INVALID_INPUT_OPTIONS_FORMAT: 'Interface: container create - Invalid type, format must be of non-empty string type.',
     INVALID_INPUT_OPTIONS_TRANSLATION: 'Interface: container create - Invalid type, translation must be of non-empty object type.',
     EMPTY_COLLECT_CUSTOM_FORMAT: 'Interface: container create - %s1 cannot be empty object.',
-
+    INVALID_ALLOWED_OPTIONS: 'Interface: collect element - Invalid options, allowedFileType must be of string array type.',
+    EMPTY_ALLOWED_OPTIONS_ARRAY: 'Interface: collect element - Invalid options value, allowedFileType array is empty.',
+    INVALID_ALLOWED_FILETYPE_ARRAY: 'Interface: collect element - Invalid options value, allowedFileType array must contain only string elements.',
   },
   warnLogs: {
     INVALID_EXPIRATION_DATE_FORMAT: 'EXPIRATION_DATE format must be in one of %s1, the format is set to default MM/YY',
@@ -284,6 +287,7 @@ const logs = {
     GET_BY_ID_DEPRECATED: 'getById is deprecated, use new get method',
     INPUT_FORMATTING_NOT_SUPPROTED: 'format or translation are not supported on %s1 element type.',
     INVALID_INPUT_TRANSLATION: 'invalid or unsupported translation provided for %s1 element type.',
+
   },
 };